refactor: 移动低级的代码到另一个文件
