+++ qb_sdk/src/lib.rs
@@ -3,8 +3,7 @@ mod web_api;
 use chrono::{DateTime, Duration, Local};
 use command::Cli;
 use log;
-use reqwest::{Client, Error};
-use serde_json::Value;
+use reqwest::Client;
 use std::collections::HashMap;
 use tokio::time::sleep;
 
@@ -46,22 +45,12 @@ pub struct Peer {
 
 impl QbClient {
     pub fn new(cli: Cli) -> Self {
-        QbClient{client: Client::new(), config: cli, last_reset_time: Local::now(), torrent_dic: HashMap::new() }
+        QbClient{client: Client::new(), config: cli,
+            last_reset_time: Local::now() - Duration::days(2), torrent_dic: HashMap::new() }
     }
 
     pub async fn wait(&self) {
         sleep(std::time::Duration::from_secs(self.config.interval)).await;
-    }
-
-    
-
-    async fn get_api_version(&self) -> Result<String, Error>
-    {
-        let resp = self.client.get(self.get_host() + "/api/v2/app/webapiVersion")
-            .send()
-            .await?;
-        let text = resp.text().await?;
-        Ok(text)
     }
 
     pub async fn ensure_api_version(&self) -> Result<(), String>
@@ -94,26 +83,6 @@ impl QbClient {
     }
 
     #[allow(non_snake_case)]
-    pub async fn reset_banned_IPs(&self) -> Result<(), String>
-    {
-        let result = self.web_api_set_preferences()
-            .form(&[("json", r#""{"banned_IPs":""}""#)])
-            .send()
-            .await;
-        let resp = match result {
-            Ok(resp) => resp,
-            Err(e) => {
-                return Err(format!("Can't reset QBittorrent IP:\n{:#?}", e));
-            }
-        };
-        if !resp.status().is_success() {
-            return Err(format!("Can't reset QBittorrent IPs:\n{:#?}", resp));
-        }
-        log::log("Reset banned IPs!");
-        Ok(())
-    }
-
-    #[allow(non_snake_case)]
     pub async fn try_reset_banned_IPs(&mut self) -> Result<(), String>
     {
         if Local::now() - self.last_reset_time > Duration::days(1) {
@@ -125,114 +94,37 @@ impl QbClient {
 
     pub async fn record_and_ban_peers(&mut self) -> Result<(), String>
     {
-        // 获取所有 torrent 的信息
-        let resp = match self.web_api_get_torrents_info().send().await {
-            Ok(resp) => resp,
-            Err(e) => return Err(format!("Can't get QBittorrent torrents info:\n{:#?}", e))
-        };
-        if !resp.status().is_success() {
-            return Err(format!("Can't get QBittorrent torrents info:\n{:#?}", resp));
-        }
-        let content = match resp.text().await {
-            Ok(t) => t,
-            Err(e) => return Err(format!("Can't get QBittorrent torrents info:\n{:#?}", e)),
-        };
-        let json_value: Value = match serde_json::from_str(&content){
-            Ok(v) => v,
-            Err(e) => return Err(format!("Can't get QBittorrent torrents info:\n{:#?}", e))
-        };
-        let hash_array = match json_value.as_array() {
-            Some(v) => v,
-            None => return Err("Can't get QBittorrent torrents info".to_string()),
-        };
-        let torrent_array: Vec<_>= hash_array
-            .iter().filter_map(|p| {
-            let hash = match p["hash"].as_str() {
-                Some(v) => v,
-                None => return None
-            };
-            let size = match p["total_size"].as_u64() {
-                None => return None,
-                Some(v) => v,
-            };
-            Some(Torrent{hash: String::from(hash), size, peer_dic: HashMap::new()})
-        })
-            .collect();
-        {
-            // 记录 torrent 的信息
-            let torrent_dic = &mut self.torrent_dic;
-            for torrent in torrent_array {
-                match torrent_dic.get(torrent.hash.as_str()) {
-                    None => {
-                        torrent_dic.insert(String::from(torrent.hash.as_str()), torrent.clone());
-                    }
-                    Some(_) => {}
+        let torrent_array_from_qb = self.get_torrents().await?;
+        for torrent in torrent_array_from_qb {
+            match self.torrent_dic.get(torrent.hash.as_str()) {
+                None => {
+                    self.torrent_dic.insert(String::from(torrent.hash.as_str()), torrent.clone());
                 }
+                Some(_) => {}
             }
         }
         // hsah, ip, info
-        let mut torrent_ip_peer: HashMap<String, HashMap<String, Peer>> = HashMap::with_capacity(self.torrent_dic.len());
+        let mut torrent_ip_peer_from_qb: HashMap<String, HashMap<String, Peer>> = HashMap::with_capacity(self.torrent_dic.len());
         {
             let torrent_dic = &self.torrent_dic;
             for (hash, _) in torrent_dic {
-                // 获取连接到这个 torrent 的所有 ip
-                let resp = match self.web_api_sync_torrent_peers(hash.as_str()).send().await {
-                    Ok(resp) => resp,
-                    Err(e) => return Err(format!("Can't get QBittorrent torrents info: {}\n{}", hash, e)),
-                };
-                if !resp.status().is_success() {
-                    return Err(format!("Can't get QBittorrent torrents info: {}\n{}", hash, resp.status()))
-                }
-                let content = match resp.text().await {
-                    Ok(t) => t,
-                    Err(e) => return Err(format!("Can't get QBittorrent torrents info: {}\n{}", hash, e)),
-                };
-                let json_value: Value = match serde_json::from_str(&content){
-                    Ok(v) => v,
-                    Err(e) => return Err(format!("Can't get QBittorrent torrents info: {}\n{}", hash, e)),
-                };
-                let json_value = match json_value["peers"].as_object() {
-                    Some(v) => v,
-                    None => return Err(format!("Can't get QBittorrent torrents info: {}", hash)),
-                };
-                let hash_peers : HashMap<String, Peer> = json_value.iter().filter_map(|(k, v)| {
-                    let ip = match v["ip"].as_str() {
-                        Some(v) => v,
-                        None => return None,
-                    };
-                    let port = match v["port"].as_i64() {
-                        Some(v) => v,
-                        None => return None,
-                    } as u16;
-                    let uploaded = match v["uploaded"].as_u64() {
-                        Some(v) => v,
-                        None => return None,
-                    };
-                    let progress = match v["progress"].as_f64() {
-                        Some(v) => v,
-                        None => return None,
-                    };
-                    let client = match v["client"].as_str() {
-                        Some(v) => v,
-                        None => return None,
-                    };
-                    Some((String::from(k), Peer{ip: String::from(ip), port, uploaded, progress, client: String::from(client), }))
-                }).collect();
-                torrent_ip_peer.insert(String::from(hash), hash_peers);
+                let hash_peers = self.get_peers(hash.as_str()).await?;
+                torrent_ip_peer_from_qb.insert(String::from(hash), hash_peers);
             }
         }
 
         // 移除没有出现的 peer
         for (_, torrent) in self.torrent_dic.iter_mut() {
-            let self_ip_ports = torrent.peer_dic.iter().filter_map(|(k, _)| {
+            let ip_ports_from_self = torrent.peer_dic.iter().filter_map(|(k, _)| {
                 Some(String::from(k))
             }).collect::<Vec<String>>();
-            let torrent_from_torrent_ip_peer = match torrent_ip_peer.get(torrent.hash.as_str()) {
+            let torrent_from_torrent_ip_peer_from_qb =
+                match torrent_ip_peer_from_qb.get(torrent.hash.as_str()) {
                 Some(v) => v,
                 None => continue,
             };
-            for ip_port in self_ip_ports {
-                if !torrent_from_torrent_ip_peer.contains_key(ip_port.as_str()) {
+            for ip_port in ip_ports_from_self {
+                if !torrent_from_torrent_ip_peer_from_qb.contains_key(ip_port.as_str()) {
                     torrent.peer_dic.remove(ip_port.as_str());
                 }
             }
@@ -240,7 +132,7 @@ impl QbClient {
 
         let mut ban_peers: Vec<String> = vec![];
         // 更新 peer 信息，并判断是否 ban
-        for (hash, peers) in torrent_ip_peer.iter() {
+        for (hash, peers) in torrent_ip_peer_from_qb.iter() {
             let torrent_size =  *&self.torrent_dic[hash.as_str()].size;
             let old_torrent = match self.torrent_dic.get_mut(hash.as_str()) {
                 None => {
@@ -278,7 +170,6 @@ impl QbClient {
         if !resp.status().is_success() {
             return Err(format!("Can't get QBittorrent torrents info:\n{}", resp.status()))
         }
-
         Ok(())
     }
     fn judge_banned(old: &Peer, new: &Peer, torrent_size: u64) -> bool
